+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-06T05:48:57.514Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-06T08:42:44.600Z")
     r"""The date and time the resource was last updated"""