+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-06T05:48:57.124Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-06T08:42:44.236Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")