+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9BVCC53G0BYFS6580A88F8Z"
+        "01K9C5AJ8DPYHEARFY76C96CFX"
     )
     r"""The unique identifier of the data source"""