+++ packages/orq-rc/src/orq_ai_sdk/models/listbudgetsop.py
@@ -250,7 +250,7 @@ class ListBudgetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-06T05:48:57.514Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-06T08:42:44.600Z")
     r"""The date and time the resource was last updated"""