+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9BVCC4ZE88JXEW1E90RM5Q0"
+        "01K9C5AJ892WY8WCR2Q87A4J4Z"
     )
     r"""The unique identifier of the data source"""