+++ packages/orq-rc/RELEASES.md
@@ -3869,3 +3869,13 @@ Based on:
 - [python v4.0.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.11 - packages/orq-rc
+
+## 2025-11-06 08:45:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.649.0 (2.739.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.12 - packages/orq-rc